does anyone read these?
born to code the wode and the smode for the dodes who implode when coded into the matrix which is in itself an illusion even once realized that it doesn't exist but still despite  that people still call it a miracle when they can escape it when really the solution was infront of them the whole time but in reality when you think about the sybolism of the film you start to see that it's not about a robotic dystopia with humanity all used to power the mechines even with the true end saying that the martriix has layers and they were in it the whole time even when they thought they were in the real world. now this is a revelation that shocked the audidence and nobody expected for it to be like this as the established theme was them going from the fake world to the real world. and that's my opinion on hotdogs thanks for reading!